changed so that app now get the path to the "files" directory as input argument. also refactored the package because of a typo
