+++ src/main/java/com/e/IndexedWord.java
@@ -1,6 +1,5 @@
-package com.zombimike;
+package com.e;
 
-import java.io.Serializable;
 import java.util.*;
 import java.util.stream.Stream;
