+++ src/test/java/com/e/AppTest.java
@@ -1,4 +1,4 @@
-package com.zombimike;
+package com.e;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
